requirements.txt
# ── Core Flask ─────────────────────────────────────────────────────────────────
Flask==3.1.2
Werkzeug==3.1.4
blinker==1.9.0
click==8.3.1
itsdangerous==2.2.0
Jinja2==3.1.6

# ── TensorFlow / Keras ─────────────────────────────────────────────────────────
# Keras 3 is standalone — import as 'keras', NOT 'tensorflow.keras'
tensorflow==2.20.0
keras==3.13.2

# ── NumPy (MUST be <2 — opencv, scikit-learn, scipy all require it) ────────────
numpy==1.26.4

# ── Image processing ───────────────────────────────────────────────────────────
Pillow==10.0.0
opencv-python-headless==4.8.0.76

# ── NLP / Chatbot ──────────────────────────────────────────────────────────────
torch==2.1.2
sentence-transformers==2.3.1
transformers==4.35.2
huggingface-hub==0.19.4
scikit-learn==1.3.2
scipy==1.11.4

# ── Supporting packages ────────────────────────────────────────────────────────
h5py==3.11.0
tqdm==4.66.1
requests==2.31.0
six==1.17.0